resolve ticket #238 Component-Layer now supports true nested transactions/savepoints including test. Changed TZConnection.StartTransaction to a function which returns the txn level. for the docs of Dbc/component.StartTransaction: Integer behavior:
Starts transaction support or saves the current transaction.
If the connection is closed, the connection will be opened(dbc only).
If a transaction is underway a nested transaction or a savepoint will be spawned.
While the tranaction(s) is/are underway the AutoCommit property is set to False.
Ending up the transaction with a commit/rollback the autocommit property will be restored if changing the autocommit mode was triggered by a starttransaction call.
Returns the current txn-level. 1 means a transaction was started.
2 means the transaction was saved. 3 means the previous savepoint got saved too and so on